==> 10-Predict_COLLtoCOLL.R <==
cat("Modelling data...\n")

# dir.create(paste0(odir, "/predictions"), showWarnings = FALSE, recursive = TRUE)
dir.create(paste0(odir, "/predictions/COLLtoCOLL"), showWarnings = FALSE, recursive = TRUE)

## if import, model phenos, mnodel genos scripts are skipped, load data here
# id_pheno<-read.table(paste0(odir, "/phenos_modelled/rownames_phenos.txt"))
## problem with duplicated rowname that I do not understand, use id_pheno to replace them (saved with phenos)
phenos<-read.table(paste0(odir, "/phenos_modelled/BLUPs_PC1_PC2_for_pred.txt"), h=T)
summary(phenos)
# phenos<-phenos[,-1]
genos_ready=readRDS(paste0(idir, "/genos_imputed_for_pred.rds"))
## clusters from DAPC analysis
clusters<-read.table(paste0(odir, "/genos_modelled/assignments_COLL_DAPC.txt"), h=T)
cluster_fams<-read.table(paste0(odir, "/genos_modelled/assignements_families.txt"), h=T)

# Predict each family with the collection
## Now we need phenos and genos_ready files
## reinitialise order of IDs
ids<-intersect(rownames(phenos), rownames(genos_ready)) %>% sort
NbID<-length(ids)
cat("Number of common genos/phenos: \n")
print(NbID)
phenos<-phenos[ids,]
genos_ready<-genos_ready[ids,]
nrow(phenos) == nrow(genos_ready)
## simple rrBLUP

## list of IDs in collection and family names
families<-c("FjDe", "FjPi", "FjPL", "GDFj", "GaPi", "GaPL")
NbFAM<-length(families)
WhichCOL<-c(1:NbID)[-c(lapply(families, function(x) grep(x, ids) ) %>% unlist)]
summary(rownames(genos_ready)[WhichCOL] == rownames(phenos)[WhichCOL]) ## 259 with same names

traits=colnames(phenos)

## help on 5-fold design: https://stats.stackexchange.com/questions/61090/how-to-split-a-data-set-to-do-10-fold-cross-validation
nREPs=100
accuracy<-data.frame(Rep=rep(1:nREPs, length(traits)), trait=lapply(traits, function(x) rep(x, nREPs)) %>% unlist,
                     accuracy_Pearson=NA, accuracy_Spearman)
count=0
for (trait in traits) {
# for (trait in c("PC1","PC2")) {
  for (REP in 1:nREPs) {
    folds <- cut(seq(1,length(WhichCOL)),breaks=5,labels=FALSE) ## create folds
    ids_shuffle<-sample(WhichCOL) ## randomize ids
    fold_acc<-c()
    for(i in 1:5){
      #Segement your data by fold using the which() function 
      WhichVS <- ids_shuffle[which(folds==i) ]
      WhichTS <- ids_shuffle[-which(folds==i) ]
      res <- mixed.solve(y=phenos[WhichTS,trait],Z=genos_ready[WhichTS,])
      Y_VS_pred<- as.vector(genos_ready[WhichVS,] %*% as.matrix(res$u))
      fold_acc_P<-append(fold_acc,cor(phenos[WhichVS, trait], Y_VS_pred, method="pearson"))
      fold_acc_S<-append(fold_acc,cor(phenos[WhichVS, trait], Y_VS_pred, method="spearman"))
      rm(WhichVS, WhichTS)
    }
    count=count+1
    accuracy[count,"accuracy_Pearson"]<-mean(fold_acc_P)
    accuracy[count,"accuracy_Spearman"]<-mean(fold_acc_S)
    print(c(mean(fold_acc_P), mean(fold_acc_S)))
  }
}    
write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/accuracy_5fold_rrBLUP_pearson_spearman.txt"), quote=F, sep="\t")
# write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/accuracy_5fold_rrBLUP.txt"), quote=F, sep="\t")

# write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/accuracy_5fold_rrBLUP_PC12_new_values.txt"), quote=F, sep="\t")

# boxplot
accuracy<-read.table(paste0(odir, "/predictions/COLLtoCOLL/accuracy_5fold_rrBLUP.txt"))

png(file=paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_5fold_", nreps, "_reps.png"), height=500, width=800)
ggplot(accuracy,aes( y=accuracy, x=trait))+
  geom_boxplot()+
  labs(x="Trait", title="Predictions rrBLUP 100 CVs 5-fold", y="Predictive ability")+
  theme(axis.text.x=element_text(angle = 45,  hjust = 1))+ 
  scale_y_continuous(limits = c(0, 1))
dev.off()

## add cluster effect
## try with clusters (as fix effect)
head(clusters)
clusters<-as.data.frame(clusters)
clusters$Cluster<-as.factor(clusters$Cluster)
clusters.mat<-class.ind(clusters$Cluster)
rownames(clusters.mat)<-clusters$Name
summary(rownames(clusters.mat)== rownames(genos_ready)[WhichCOL])
nREPs=100
accuracy<-data.frame(Rep=rep(1:nREPs, length(traits)), trait=lapply(traits, function(x) rep(x, nREPs)) %>% unlist,accuracy=NA)
count=0

# traits=c("PC1", "PC2")
# accuracy<-data.frame(Rep=rep(1:nREPs, length(traits)), trait=lapply(traits, function(x) rep(x, nREPs)) %>% unlist,accuracy=NA)

for (trait in traits) {
  print(trait)
  for (REP in 1:nREPs) {
    folds <- cut(seq(1,length(WhichCOL)),breaks=5,labels=FALSE) ## create folds
    ids_shuffle<-sample(WhichCOL) ## randomize ids
    fold_acc<-c()
    for(i in 1:5){
      #Segement your data by fold using the which() function 
      WhichVS <- ids_shuffle[which(folds==i) ]
      WhichTS <- ids_shuffle[-which(folds==i) ]
      res <- mixed.solve(y=phenos[WhichTS,trait],Z=genos_ready[WhichTS,], X=clusters.mat[ids[WhichTS],])
      Y_VS_pred<- as.vector(genos_ready[WhichVS,] %*% as.matrix(res$u)  + clusters.mat[ids[WhichVS],] %*% res$beta)
      fold_acc<-append(fold_acc,cor(phenos[WhichVS, trait], Y_VS_pred))
      rm(WhichVS, WhichTS)
    }
    count=count+1
    accuracy[count,"accuracy"]<-mean(fold_acc)
    print(mean(fold_acc))
    
  }
  write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/rrBLUPs_5fold_6clusters_fix.txt"), quote=F, sep="\t")

  
}

summary(accuracy)
accuracy<-read.table(paste0(odir, "/predictions/COLLtoCOLL/rrBLUPs_5fold_6clusters_fix.txt"))

## architecture of traits
png(file=paste0(odir, "/predictions/COLLtoCOLL/architecture_traits.png"), width=1500, height=600)
par(mfrow=c(3,5))
mynorm<-function(x) (x - min(x))/(max(x) - min(x))
for (trait in traits) {
      #Segement your data by fold using the which() function 
      data<- mynorm(phenos[,trait])
      res <- mixed.solve(y=data,Z=genos_ready)
      print(hist(res$u, main=paste0(trait, "\n", "(sd:", round(sd(res$u), digits=4),")"),
                 xlab="Effect of marker", nclass=20))
} 
dev.off()

############### not used for paper below ####################
# 
# ## compare with Monte Carlo
# nreps=100
# accuracy<-data.frame(Rep=rep(1:nreps, length(traits)), trait=lapply(traits, function(x) rep(x, nreps)) %>% unlist,accuracy=NA)
# count=0
# for (trait in traits) {
#   for (rep in 1:nreps) {
#     WhichTS<-sample(WhichCOL, round(length(WhichCOL))*4/5)
#     res <- mixed.solve(y=phenos[WhichTS,trait],Z=genos_ready[WhichTS,])
#     WhichVS<-setdiff(WhichCOL, WhichTS)
#     Y_VS_pred<- as.vector(genos_ready[WhichVS,] %*% as.matrix(res$u))
#     acc<-cor(phenos[WhichVS, trait], Y_VS_pred, use="na.or.complete")
#     count=count+1
#     accuracy[count,"accuracy"]<-acc
#     # print(c(trait, rep, i, count))
#     rm(WhichVS, WhichTS)
#   }
# }    
# 
# write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_MC_", nreps, "_reps.txt"), quote=F, sep="\t")
# 
# png(file=paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_MC_", nreps, "_reps.png"), height=500, width=800)
# ggplot(accuracy,aes( y=accuracy, x=trait))+
#   geom_boxplot()+
#   labs(x="Trait", title="Predictions rrBLUP 100 Monte Carlo 1/5", y="Predictive ability")+
#   theme(axis.text.x=element_text(angle = 45,  hjust = 1))+
#   scale_y_continuous(limits = c(0, 1))
# dev.off()
# 
# ## validate within same year
# head(phenos_raw)
# phenos_col<-phenos_raw[which(phenos_raw[,"Family"]=="Collection"),] %>% droplevels
# phenos_col<-phenos_col[order(phenos_col$Name),]
# phenos_col$Year<-as.factor(phenos_col$Year)
# head(phenos_col)
# nbY<-length(levels(phenos_col$Year))
# traits<-colnames(phenos_col)[-c(1:7)]
# accuracy<- data.frame(Rep=rep(1:nreps, length(traits)*nbY), 
#                      Year=lapply(levels(phenos_col$Year), function(x) rep(x, nreps*length(traits)))%>% unlist,
#                      trait=lapply(traits, function(x) rep(x, nreps)) %>% unlist %>% rep(. ,nbY),
#                      accuracy=NA )
# nreps=100
# count=0
# for (Y in levels(phenos_col$Year)) {
#   data<-phenos_col[which(phenos_col$Year==Y),] %>% droplevels
#   for (trait in traits) {
#     NonMissing<-which(!(is.na(data[,trait]))) 
#     pheno<-data[NonMissing,] %>% droplevels 
#     pheno<- aggregate(x=pheno[,trait],by=list(pheno$Name),FUN= mean)
#     rownames(pheno)<-pheno$Group.1
#     ids<-intersect(rownames(genos_ready), rownames(pheno)) %>% sort
#     pheno<-pheno[ids,]
#     WhichCOL<-which(rownames(genos_ready) %in% ids)
#     summary(rownames(genos_ready)[WhichCOL] == rownames(pheno))
#     for (rep in 1:nreps) {
#       folds <- cut(seq(1,length(WhichCOL)),breaks=5,labels=FALSE) ## create folds
#       ids_shuffle<-sample(WhichCOL) ## randomize ids
#       fold_acc<-c()
#       for(i in 1:5){
#         #Segment your data by fold using the which() function 
#         WhichVS <- ids_shuffle[which(folds==i) ]
#         WhichTS <- ids_shuffle[-which(folds==i) ]
#         res <- mixed.solve(y=pheno[rownames(genos_ready)[WhichTS],"x"],Z=genos_ready[WhichTS,])
#         Y_VS_pred<- as.vector(genos_ready[WhichVS,] %*% as.matrix(res$u))
#         fold_acc<-append(fold_acc,cor(pheno[rownames(genos_ready)[WhichVS],"x"], Y_VS_pred, use="na.or.complete"))
#         rm(WhichVS, WhichTS)
#       }
#       count=count+1
#       accuracy[count,"accuracy"]<-mean(fold_acc)
#       # print(c(trait, rep, i, count))
#     }
#   }    
# }
# write.table(accuracy, file=paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_per_year_5Fold_", nreps, "_reps.txt"),sep="\t", quote=F)
# 
# 
# accuracy<-read.table(paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_per_year_5Fold_", nreps, "_reps.png"), sep="\t", h=T)
# 
# 
# png(file=paste0(odir, "/predictions/COLLtoCOLL/COLLtoCOLL_per_year_5Fold_", nreps, "_reps.png"), height=500, width=1000)
# 
# ggplot(accuracy,aes( y=accuracy, x=trait))+
#   geom_boxplot()+
#   facet_grid(~Year) +
#   labs(x="Trait", title="Predictions rrBLUP 100 CVs 5-fold", y="Predictive ability")+
#   theme(axis.text.x=element_text(angle = 45,  hjust = 1))+
#   scale_y_continuous(limits = c(-0.5, 1))
# dev.off()







# Purge obsolete variables
rm()

cat("Data modelled!\n")
